+++ tests/Mongofill/Tests/Integration/Standalone/MongoCursorTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mongofill\Tests\Integration;
+namespace Mongofill\Tests\Integration\Standalone;
 
 class MongoCursorTest extends TestCase
 {
