+++ tests/Mongofill/Tests/Integration/Standalone/TestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mongofill\Tests\Integration;
+namespace Mongofill\Tests\Integration\Standalone;
 
 use PHPUnit_Framework_TestCase;
 use MongoClient;
